Fixed objective creation and automatized eye count.
Added a new boople with 7 eyes, and this brought up a bug: the way it
created missions from a dictionary (checking it's size), so it wanted 5
eyed. Fixed that with getting a random elemnt from the list of the
dictionary's keys.
Also added a simple line, which sets the CreatureLogic component's
eyeCount member automatically, so I don't have to manually adjust it
when adding a new one.